requirements.txt
tensorflow
opencv-python
numpy
matplotlib
nvidia-ml-py3
Pillow
streamlit
requests
opencv-python
tensorflow.keras
streamlit